[dpo] initialize dpo for presto
