+++ requirements.txt
@@ -10,6 +10,7 @@ deepspeed
 torch_geometric
 rdkit>=2021.03.5
 selfies>=1.0.0
+trl>=0.4.7
 wandb
 scikit-learn
 nltk
